V4 (#107)
feat: Update buttplug4j to the v4 spec

This will only work with the Interface Central v3 release, currently in beta.

There is still know issues with InputCommand repsonses

+++ buttplug4j/src/main/java/io/github/blackspherefollower/buttplug4j/protocol/messages/Ping.java
@@ -10,7 +10,7 @@ public final class Ping extends ButtplugMessage {
         super(ButtplugConsts.DEFAULT_MSG_ID);
     }
 
-    public Ping(final long id) {
+    public Ping(final int id) {
         super(id);
     }
 }
