+++ buttplug4j/src/main/java/io/github/blackspherefollower/buttplug4j/protocol/messages/StartScanning.java
@@ -10,7 +10,7 @@ public final class StartScanning extends ButtplugMessage {
         super(ButtplugConsts.DEFAULT_MSG_ID);
     }
 
-    public StartScanning(final long id) {
+    public StartScanning(final int id) {
         super(id);
     }
 }
